Moved exceptions to api package.
